+++ src/Paciente.java
@@ -1,6 +1,6 @@
 public class Paciente {
     
-    private Integer id_paciente;
+    private Long id_paciente;
     private String nome;
     private String dt_nascimento;
     private String cpf;
@@ -8,7 +8,7 @@ public class Paciente {
     private String motivo;
     
     //Construtor
-    public Paciente(Integer id_paciente, String nome, String dt_nascimento, String cpf, String idade, String motivo) {
+    public Paciente(Long id_paciente, String nome, String dt_nascimento, String cpf, String idade, String motivo) {
         this.id_paciente = id_paciente;
         this.nome = nome;
         this.dt_nascimento = dt_nascimento;
@@ -18,11 +18,11 @@ public class Paciente {
     }
 
     // Getters e Setters
-    public Integer getId_paciente() {
+    public Long getId_paciente() {
         return id_paciente;
     }
 
-    public void setId_paciente(Integer id_paciente) {
+    public void setId_paciente(Long id_paciente) {
         this.id_paciente = id_paciente;
     }
