+++ src/Perfil.java
@@ -1,22 +1,22 @@
 public class Perfil {
-    private Integer id_perfil;
+    private Long id_perfil;
     private String nome;
     private String descricao;
 
     //Construtor
-    public Perfil(Integer id_perfil, String nome, String descricao) {
+    public Perfil(Long id_perfil, String nome, String descricao) {
         this.id_perfil = id_perfil;
         this.nome = nome;
         this.descricao = descricao;
     }
 
     // Getters e Setters
-    public Integer getId() {
+    public Long getId() {
         return id_perfil;
     }
 
 
-    public void setId(Integer id_perfil) {
+    public void setId(Long id_perfil) {
         this.id_perfil = id_perfil;
     }
